Change hat state to UInt8
Return hat state enum value instead of UInt8 from get_joystick_hats.

+++ src/glfw/input/joystick_hat_state.cr
@@ -1,6 +1,6 @@
 lib LibGLFW
   @[Flags]
-  enum JoystickHatState : Int32
+  enum JoystickHatState : UInt8
     Centered  = 0
     Up        = 1
     Right     = 2
